### 2025-06-20 #8434 @editorialbot
[PRE REVIEW]: OperationsResearchModels.jl: A Julia package for operations research models
**Submitting author:** @jbytecode (<a href="http://orcid.org/0000-0002-9402-1982">Mehmet Hakan Satman</a>)
**Repository:** https://github.com/jbytecode/OperationsResearchModels.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.5
