### 2025-06-20 47ac2a7
remove display in test of shortestpath
