### 2025-06-20 9e36c4d
change the order of tests
