+++ test/runtests.jl
@@ -3,13 +3,13 @@ using OperationsResearchModels
 using OperationsResearchModels.Simplex
 
 
+include("testjump.jl")
+include("testassignment.jl")
 include("testtransportation.jl")
 include("testrandomkeyga.jl")
 include("testtravelingsalesman.jl")
 include("testjohnsons.jl")
 include("testutility.jl")
-include("testjump.jl")
-include("testassignment.jl")
 include("testshortestpath.jl")
 include("testmaximumflow.jl")
 include("testgame.jl")
