### 2025-06-20 934d1e8
add 100x100 test for the assignment problem
