+++ test/testassignment.jl
@@ -22,4 +22,28 @@
             1 0 0
         ]
     end
+
+
+    @testset "Relatively big Assignment Problem (100x100)" begin
+
+        n = 100
+
+        mat = rand(10:1000, n, n)
+
+        # Ensure that the ith row has a 1 in the ith column
+        for i in 1:n
+            mat[i, i] = 1
+        end
+
+        a = AssignmentProblem(mat)
+        result::AssignmentResult = solve(a)
+
+        @test result.problem isa AssignmentProblem
+        @test result.cost == n
+        
+        for i in 1:n
+            @test result.solution[i, i] == 1
+        end 
+
+    end
 end
