### 2025-06-20 fe85bce
replace Array{T,1} and Array{T, 2} with Vector{T} and Matrix{T}, respectively
