+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make initial solution optional in Transportation problems 
 - Add tests for the shortest problem problem
 - Add a 100x100 test problem for assignment
+- Replace `Array{Type, 1}` and `Array{Type, 2}` with `Vector{T}` and `Matrix{Type}`, respectively
 
 
 ### 0.2.5 
