### 2025-06-22 218a8cf
add test for shortest path problem (two way connections)
