+++ test/runtests.jl
@@ -12,6 +12,7 @@ include("testjohnsons.jl")
 include("testutility.jl")
 include("testshortestpath.jl")
 include("testmaximumflow.jl")
+include("testminimumcostflow.jl")
 include("testgame.jl")
 include("testmst.jl")
 include("testpmedian.jl")
