+++ src/maximumflow.jl
@@ -6,8 +6,8 @@ using JuMP, HiGHS
 import ..OperationsResearchModels: solve
 
 
-export MaximumFlowProblem
-export MaximumFlowResult
+export MaximumFlowProblem, MaximumFlowResult
+export MinimumCostFlowProblem, MinimumCostFlowResult
 
 
 struct MaximumFlowResult
@@ -19,6 +19,66 @@ struct MaximumFlowProblem
     connections::Vector{Connection}
 end
 
+struct MinimumCostFlowProblem
+    connections::Vector{Connection}
+    costs::Vector{Connection}
+end
+
+struct MinimumCostFlowResult 
+    path::Vector{Connection}
+    cost::Float64
+end 
+
+
+function leftexpressions(x::Matrix{JuMP.VariableRef}, node::Int64, nodes::Vector{Connection}, model)
+    lst = []
+    for conn in nodes
+        if conn.to == node
+            push!(lst, conn)
+        end
+    end
+    if length(lst) == 0
+        return :f
+    end
+    expr = @expression(model, 0)
+    for i = eachindex(lst)
+        expr += x[lst[i].from, lst[i].to]
+    end
+    return expr
+end
+
+function rightexpressions(x::Matrix{JuMP.VariableRef}, node::Int64, nodes::Vector{Connection}, model)
+    lst = []
+    for conn in nodes
+        if conn.from == node
+            push!(lst, conn)
+        end
+    end
+    if length(lst) == 0
+        return :f
+    end
+    expr = @expression(model, 0)
+    for i = eachindex(lst)
+        expr += x[lst[i].from, lst[i].to]
+    end
+    return expr
+end
+
+
+function hassameorder(a::Vector{Connection}, b::Vector{Connection})::Bool
+
+    if length(a) != length(b)
+        return false
+    end
+
+    for i = 1:length(a)
+        if a[i].from != b[i].from || a[i].to != b[i].to
+            return false
+        end
+    end
+
+    return true
+end
 
 """
 
@@ -65,44 +125,9 @@ julia> result.flow
 7.0
 ```
 """
-function solve(problem::MaximumFlowProblem)
+function solve(problem::MaximumFlowProblem)::MaximumFlowResult
 
     cns = problem.connections
-
-    function leftexpressions(node::Int64, nodes::Vector{Connection}, model)
-        lst = []
-        for conn in nodes
-            if conn.to == node
-                push!(lst, conn)
-            end
-        end
-        if length(lst) == 0
-            return :f
-        end
-        expr = @expression(model, 0)
-        for i = eachindex(lst)
-            expr += x[lst[i].from, lst[i].to]
-        end
-        return expr
-    end
-
-    function rightexpressions(node::Int64, nodes::Vector{Connection}, model)
-        lst = []
-        for conn in nodes
-            if conn.from == node
-                push!(lst, conn)
-            end
-        end
-        if length(lst) == 0
-            return :f
-        end
-        expr = @expression(model, 0)
-        for i = eachindex(lst)
-            expr += x[lst[i].from, lst[i].to]
-        end
-        return expr
-    end
-
 
     model = Model(HiGHS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
@@ -122,8 +147,8 @@ function solve(problem::MaximumFlowProblem)
 
     # Constraints 
     for nextnode in mynodes
-        leftexpr = leftexpressions(nextnode, cns, model)
-        rightexpr = rightexpressions(nextnode, cns, model)
+        leftexpr = leftexpressions(x, nextnode, cns, model)
+        rightexpr = rightexpressions(x, nextnode, cns, model)
         if leftexpr == :f
[…]
