### 2025-07-04 05b9e55
add docs for minimum cost flow
