+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 JuMP [@JuMP] provides an excellent interface and macros for uniformly accessing optimizer functionality. Any mathematical optimization problem can be assembled with three core components: the objective function (`@objective`), variable definitions (`@variable`), and constraints (`@constraints`). The researcher's role is to formulate the original problem as a mathematical optimization problem and then translate it using JuMP's macros.
 
-OperationsResearchModels.jl streamlines the model translation stage by automatically constructing mathematical problems from problem-specific input data. Its extensive functionality encompasses a significant portion of the Operations Research domain. This includes, but is not limited to: the linear transportation problem, the assignment problem, the classical knapsack problem, various network models (Shortest Path, Maximum Flow, Minimum Spanning Tree), project management techniques (CPM and PERT), the uncapacitated p-median problem for location selection, Johnson's Rule for flow-shop scheduling, a genetic algorithm for scheduling problems intractable by Johnson's Rule, a zero-sum game solver, and a Simplex solver for real-valued decision variables.
+OperationsResearchModels.jl streamlines the model translation stage by automatically constructing mathematical problems from problem-specific input data. Its extensive functionality encompasses a significant portion of the Operations Research domain. This includes, but is not limited to: the linear transportation problem, the assignment problem, the classical knapsack problem, various network models (Shortest Path, Maximum Flow, Minimum Cost-Flow, Minimum Spanning Tree), project management techniques (CPM and PERT), the uncapacitated p-median problem for location selection, Johnson's Rule for flow-shop scheduling, a genetic algorithm for scheduling problems intractable by Johnson's Rule, a zero-sum game solver, and a Simplex solver for real-valued decision variables.
 
 Although the majority of computations are performed by the HiGHS optimizer [@HiGHS] through JuMP, OperationsResearchModels.jl incorporates dedicated, hand-coded Simplex routines. These routines serve a valuable pedagogical purpose in Operations Research curricula, enabling users to observe and verify the detailed, step-by-step calculations involved in solving linear programming problems.
 
