### 2025-07-04 0ba8fcc
add test from a textbook to Minimum Cost Flow
