### 2025-07-04 a7740fb
relocate shared network functions in a separate file
