### 2025-07-06 d2d9c4d
return type declation for network utility functions
