+++ src/assignment.jl
@@ -5,8 +5,10 @@ using JuMP, HiGHS
 
 export AssignmentProblem
 export AssignmentResult
+export isbalanced
+export balance
 
-import ..OperationsResearchModels: solve
+import ..OperationsResearchModels: solve, isbalanced, balance
 
 
 """
@@ -23,6 +25,8 @@ The `AssignmentProblem` struct represents an assignment problem with a cost matr
 struct AssignmentProblem{T<:Real}
     costs::Matrix{T}
 end
+
+
 
 
 
@@ -44,6 +48,56 @@ struct AssignmentResult
     problem::AssignmentProblem
     solution::Matrix
     cost::Real
+end
+
+
+"""
+    isbalanced(problem::AssignmentProblem)::Bool
+
+# Description
+
+Checks if the assignment problem is balanced, meaning the number of workers equals the number of tasks.
+
+# Arguments
+
+- `problem::AssignmentProblem`: The assignment problem to check.
+
+# Returns
+
+- `Bool`: Returns `true` if the problem is balanced, otherwise `false`.
+"""
+function isbalanced(problem::AssignmentProblem)::Bool
+    n, p = size(problem.costs)
+    return n == p
+end
+
+
+"""
+    balance(problem::AssignmentProblem)::AssignmentProblem
+
+# Description
+
+Balances the assignment problem by adding dummy rows or columns if necessary.
+
+# Arguments
+
+- `problem::AssignmentProblem`: The assignment problem to balance.
+
+# Returns
+
+- `AssignmentProblem`: A new `AssignmentProblem` instance with balanced costs.
+"""
+function balance(problem::AssignmentProblem)::AssignmentProblem
+    n, p = size(problem.costs)
+    if n == p
+        return problem
+    elseif n < p
+        new_costs = vcat(problem.costs, zeros(p-n, p))
+        return AssignmentProblem(new_costs)
+    else
+        new_costs = hcat(problem.costs, zeros(n, n-p))
+        return AssignmentProblem(new_costs)
+    end
 end
 
 
@@ -90,10 +144,15 @@ function solve(a::AssignmentProblem)::AssignmentResult
     MOI.set(model, MOI.Silent(), true)
 
     n, p = size(a.costs)
-    @assert n == p
+    problem = a
+
+    if n != p 
+        problem  = balance(a)
+        n, p = size(problem.costs)
+    end
 
     @variable(model, x[1:n, 1:p], Bin)
-    @objective(model, Min, sum(a.costs .* x))
+    @objective(model, Min, sum(problem.costs .* x))
 
     @constraint(model, sum(x[1:n, j] for j = 1:p) .== 1.0)
     @constraint(model, sum(x[i, 1:p] for i = 1:n) .== 1.0)
