+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add tests for the shortest problem problem
 - Add a 100x100 test problem for assignment
 - Replace `Array{Type, 1}` and `Array{Type, 2}` with `Vector{T}` and `Matrix{Type}`, respectively
+- Implement `isbalanced` and `balance` for Assignment Problem
+- Solve unbalanced assignment problems
 
 
 ### 0.2.5 
