### 2025-07-07 335aaac
fix test result in case of alternative solutions
