### 2025-07-09 fe925f6
fix typo in CHANGELOG
