+++ versions.js
@@ -1,5 +1,7 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "dev";
-var DOCUMENTER_STABLE = "dev";
+var DOCUMENTER_NEWEST = "v0.2.6";
+var DOCUMENTER_STABLE = "stable";
