### 2025-08-12 d5817a3
(#1) Copy contents of README.md to /doc/src/index.md
