+++ docs/src/index.md
@@ -1,9 +1,63 @@
-## Operations Research Models
+[![Doc](https://img.shields.io/badge/docs-dev-blue.svg)](https://jbytecode.github.io/OperationsResearchModels.jl/dev/)
+[![status](https://joss.theoj.org/papers/0f312c63cdc147d43c2b899478461769/status.svg)](https://joss.theoj.org/papers/0f312c63cdc147d43c2b899478461769)
+
+# OperationsResearchModels.jl
+
+A package for Operations Research problems.
 
 
-The OperationsResearchModels package includes basic Operations Research subjects such as Transportation Problem, Assignment Problem, Minimum Spanning Tree, Shortest Path, Maximum Flow, 
-and p-medians method for selecting location of facilities. 
+# Installation 
 
-Package content is incrementally updated.
+```julia
+julia> ]
+(@v1.xx) pkg> add OperationsResearchModels
+```
 
-Please refer the Algorithms section for the detailed documentation. 
+or 
+
+```julia
+julia> using Pkg
+julia> Pkg.add("OperationsResearchModels")
+```
+
+If you want to install latest source that is not registered yet, you can 
+
+```julia
+julia> ]
+(@1.xx) pkg> add https://github.com/jbytecode/OperationsResearchModels.jl
+```
+
+but it's not recommended.
+
+# Documentation 
+
+Please visit [The Documentation Page](https://jbytecode.github.io/OperationsResearchModels.jl/dev/). 
+This page includes both the API documentation and examples.
+
+
+# Implemented Problems and Algorithms
+
+- [Assignment Problem](https://jbytecode.github.io/OperationsResearchModels.jl/dev/assignment/)
+- [Transportation Problem](https://jbytecode.github.io/OperationsResearchModels.jl/dev/transportation/)
+- [The Shortest Path](https://jbytecode.github.io/OperationsResearchModels.jl/dev/network/#Shortest-Path)
+- [Maximum Flow](https://jbytecode.github.io/OperationsResearchModels.jl/dev/network/#Maximum-Flow)
+- [Minimum Cost-Flow](https://jbytecode.github.io/OperationsResearchModels.jl/dev/network/#Minimum-Cost-Flow)
+- [Minimum Spanning Tree](https://jbytecode.github.io/OperationsResearchModels.jl/dev/network/#Minimum-Spanning-Tree)
+- [p-median for Location Selection](https://jbytecode.github.io/OperationsResearchModels.jl/dev/locationselection/)
+- [CPM - Critical Path Method](https://jbytecode.github.io/OperationsResearchModels.jl/dev/project/#CPM-(Critical-Path-Method))
+- [PERT - Project Evaluation and Review Technique](https://jbytecode.github.io/OperationsResearchModels.jl/dev/project/#PERT-(Project-Evaluation-and-Review-Technique))
+- [The Knapsack Problem](https://jbytecode.github.io/OperationsResearchModels.jl/dev/knapsack/)
+- [Johnson's Rule for Flow-Shop Scheduling](https://jbytecode.github.io/OperationsResearchModels.jl/dev/scheduling/#Johnson's-Rule-for-Flow-shop-Scheduling)
+- [Flow-shop Scheduling using Permutation Encoded Genetic Algorithms (using a Random Key Genetic Algorithm)](https://jbytecode.github.io/OperationsResearchModels.jl/dev/scheduling/#Genetic-Algorithm-for-the-problems-that-cannot-be-solved-with-using-Johnson's-Rule)
+- [Traveling Salesman with Random Key Genetic Algorithm](https://jbytecode.github.io/OperationsResearchModels.jl/dev/travelingsalesman/)
+- [Simplex Method with Real Valued Decision Variables](https://jbytecode.github.io/OperationsResearchModels.jl/dev/simplex/)
[…]
