### 2025-08-12 85e973d
(#1) Problem -> Solver -> Result order in documentations
