### 2025-08-12 0fd53ca
(#1) Add initial State of the field subsection and add citations for some related packages
