+++ paper/paper.bib
@@ -31,3 +31,19 @@
   doi={10.1007/s12532-017-0130-5},
   publisher={Springer}
 }
+
+@misc{gurobi-optimods,
+  author = {{Gurobi Optimization, LLC}},
+  title = {{Gurobi OptiMods: Data-driven APIs for common optimization tasks}},
+  howpublished = {\url{https://github.com/Gurobi/gurobi-optimods}},
+  year = {2023},
+  note = {Accessed: 2025-08-12}
+}
+
+
+@misc{Graphs2021,
+  author       = {Fairbanks, James and Besan{\c{c}}on, Mathieu and Simon, Sch{\"o}lly and Hoffiman, J{\'u}lio and Eubank, Nick and Karpinski, Stefan},
+  title        = {JuliaGraphs/Graphs.jl: an optimized graphs package for the Julia programming language},
+  year         = 2021,
+  url = {https://github.com/JuliaGraphs/Graphs.jl/}
+}
