### 2025-08-12 ed32312
(#1) Add more citations in State of the field section to address similar works
