+++ paper/paper.bib
@@ -47,3 +47,20 @@
   year         = 2021,
   url = {https://github.com/JuliaGraphs/Graphs.jl/}
 }
+
+@software{ortools,
+  title = {OR-Tools},
+  version = { v9.12 },
+  author = {Laurent Perron and Vincent Furnon},
+  organization = {Google},
+  url = {https://developers.google.com/optimization/},
+  date = { 2025-02-17 }
+}
+
+@book{kwon2019julia,
+  title={Julia Programming for Operations Research},
+  author={Kwon, Changhyun},
+  year={2019},
+  publisher={Changhyun Kwon},
+  isbn={9781798205471}
+}
