### 2025-08-25 1046ca6
(#2) Add tables for the example to make the problem more clear
