+++ paper/paper.bib
@@ -43,7 +43,7 @@
 
 @misc{Graphs2021,
   author       = {Fairbanks, James and Besan{\c{c}}on, Mathieu and Simon, Sch{\"o}lly and Hoffiman, J{\'u}lio and Eubank, Nick and Karpinski, Stefan},
-  title        = {JuliaGraphs/Graphs.jl: an optimized graphs package for the Julia programming language},
+  title        = {JuliaGraphs/Graphs.jl: an optimized graphs package for the {J}ulia programming language},
   year         = 2021,
   url = {https://github.com/JuliaGraphs/Graphs.jl/}
 }
