### 2025-08-25 c46fec5
(#2) Update Type documentation of KnapsackProblem
