### 2025-08-25 d6a47a4
(#2) Add type documentation for Connection struct
