### 2025-08-25 fccbd6d
(#2) Add type documentations for network models
