+++ docs/src/network.md
@@ -6,17 +6,49 @@
 OperationsResearchModels.Connection
 ```
 
+## ShortestPathProblem
+
+```@docs
+OperationsResearchModels.ShortestPathProblem
+```
+
 ## Shortest Path
 
 ```@docs
 OperationsResearchModels.solve(t::ShortestPathProblem)
 ```
 
+## ShortestPathResult
+
+```@docs
+OperationsResearchModels.ShortestPathResult
+```
+
+## Maximum Flow Problem
+
+```@docs
+OperationsResearchModels.MaximumFlowProblem
+```
+
+
 ## Maximum Flow
 
 ```@docs
 OperationsResearchModels.solve(t::MaximumFlowProblem)
 ```
+
+## MaximumFlowResult
+
+```@docs
+OperationsResearchModels.MaximumFlowResult
+```
+
+## MinimumCostFlowProblem 
+
+```@docs
+OperationsResearchModels.MinimumCostFlowProblem
+```
+
 
 ## Minimum Cost Flow 
 
@@ -29,8 +61,28 @@ OperationsResearchModels.solve(t::MinimumCostFlowProblem)
 OperationsResearchModels.solve(t::MinimumCostFlowProblem, flow::Float64)
 ```
 
+
+## MinimumCostFlowResult
+
+```@docs
+OperationsResearchModels.MinimumCostFlowResult
+```
+
+
+## MstProblem 
+
+```@docs
+OperationsResearchModels.MstProblem
+```
+
 ## Minimum Spanning Tree
 
 ```@docs
 OperationsResearchModels.solve(p::MstProblem)
+```
+
+## MstResult
+
+```@docs
+OperationsResearchModels.MstResult
 ```
