### 2025-08-25 16a25c8
(#2) Add documentation that states the Connection in MstProblem acts like undirected edge
