### 2025-08-25 0fd3f16
(#2) Remove `varsym` of Connection struct and set type of `value` as real
