### 2025-08-25 b9e9025
(#2) Replace `leftexpressions` and `rightexpressions` with `inflow` and `outflow`, respectively
