+++ src/maximumflow.jl
@@ -1,6 +1,6 @@
 module MaximumFlow
 
-import ..Network: Connection, nodes, start, finish, hassameorder, leftexpressions, rightexpressions
+import ..Network: Connection, nodes, start, finish, hassameorder, inflow, outflow
 
 using JuMP, HiGHS
 import ..OperationsResearchModels: solve
@@ -118,8 +118,8 @@ function solve(problem::MaximumFlowProblem)::MaximumFlowResult
 
     # Constraints 
     for nextnode in mynodes
-        leftexpr = leftexpressions(x, nextnode, cns, model)
-        rightexpr = rightexpressions(x, nextnode, cns, model)
+        leftexpr = inflow(x, nextnode, cns, model)
+        rightexpr = outflow(x, nextnode, cns, model)
         if leftexpr == :f
             @constraint(model, rightexpr == f)
         elseif rightexpr == :f
