### 2025-08-25 bda09c3
(#2) Remove `@fastmath` macro
