### 2025-08-25 d81fe67
(#2) Introduce types for traveling salesman problem and implement solve method to multiple dispatch
