+++ src/OperationsResearchModels.jl
@@ -87,7 +87,7 @@ import .Knapsack: KnapsackResult, KnapsackProblem
 import .Latex: latex
 import .Johnsons: JohnsonResult, johnsons, JohnsonException, makespan, johnsons_ga
 import .RandomKeyGA: Chromosome, run_ga
-import .TravelingSalesman: TravelinSalesmenResult, travelingsalesman
+import .TravelingSalesman: TravelingSalesmanResult, TravelingSalesmanProblem
 import .Simplex: SimplexProblem, simplexiterations, createsimplexproblem, gaussjordan
 
 export TransportationProblem, TransportationResult, balance, isbalanced, northwestcorner, leastcost
@@ -105,7 +105,7 @@ export Utility
 export latex
 export Chromosome, run_ga
 export JohnsonResult, johnsons, JohnsonException, makespan, johnsons_ga
-export TravelinSalesmenResult, travelingsalesman
+export TravelingSalesmanResult, TravelingSalesmanProblem
 export simplexiterations, SimplexProblem, createsimplexproblem, gaussjordan
 
 export JuMP, HiGHS
