+++ docs/src/travelingsalesman.md
@@ -1,5 +1,20 @@
-# Traveling Salesman
+# Traveling Salesman 
+
+## TravelingSalesmanProblem
 
 ```@docs
-OperationsResearchModels.travelingsalesman
+OperationsResearchModels.TravelingSalesmanProblem
+```
+
+## solve
+
+```@docs
+OperationsResearchModels.solve(t::TravelingSalesmanProblem)
+```
+
+
+## TravelingSalesmanResult
+
+```@docs
+OperationsResearchModels.TravelingSalesmanResult
 ```
