### 2025-08-25 2581809
(#2) Add and update documentation in Simplex module
