+++ docs/src/simplex.md
@@ -1,10 +1,24 @@
 # The Simplex Method
 
-## Simplex
+
+## createsimplexproblem
 
 ```@docs 
 OperationsResearchModels.createsimplexproblem
 ```
+
+## solve!
+
+```@docs 
+OperationsResearchModels.solve!(s::SimplexProblem)
+```
+
+## simplexiterations
+
+```@docs 
+OperationsResearchModels.simplexiterations
+```
+
 
 ## Gauss Jordan steps for matrix inversion
 
