### 2025-08-25 3e7d394
(#2) Add algorithm info in Minimal spanning tree
