### 2025-08-25 b317daf
(#2) Update examples in the network models
