### 2025-08-25 7583404
(#2) Inroduce GameProblem for game solver and implement `solve` to multiple dispatch
