### 2025-08-25 6033c11
(#2) Update documentation of the game
