+++ docs/src/game.md
@@ -7,7 +7,7 @@
 OperationsResearchModels.GameProblem
 ```
 
-## Game solver for the row player 
+## Game solver of the game matrix designed for the row player 
 
 ```@docs 
 OperationsResearchModels.solve(g::GameProblem)
