### 2025-08-25 jbytecode/OperationsResearchModels.jl#2 @jbytecode
@dourouc05 - I've tried to address all of your suggestions. Please tell me if anything is missing. The new status of the package is better. I've also triggered an `editorialbot generate pdf` command in the main issue of https://github.com/openjournals/joss-reviews/issues/8592. Thank you in advance.
