### 2025-08-26 dd75fff
(#2) Implement solve() interface for Johnson (Flow-Shop Scheduling)
