+++ docs/src/scheduling.md
@@ -1,9 +1,22 @@
 # Scheduling Problems 
 
+## Flow-Shop Problem
+
+```@docs
+OperationsResearchModels.JohnsonProblem
+```
+
+
 ## Johnson's Rule for Flow-shop Scheduling
 
 ```@docs
-OperationsResearchModels.johnsons
+OperationsResearchModels.solve(problem::JohnsonProblem)
+```
+
+## Flow-Shop Result
+
+```@docs
+OperationsResearchModels.JohnsonResult
 ```
 
 ## Genetic Algorithm for the problems that cannot be solved with using Johnson's Rule
