### 2025-08-26 782c777
(#2) Fix examples in documentation of Johnsons Rule module
