### 2025-08-26 25be985
(#2) Add type docs for CpmProblem, CpmResult, PertProblem, PertResult
