+++ docs/src/project.md
@@ -1,23 +1,64 @@
 # Project Analysis 
 
-
-## CPM (Critical Path Method)
-```@docs
-OperationsResearchModels.solve(problem::CpmProblem)
-```
-
 ### CPM Activity
+
 ```@docs
 OperationsResearchModels.CpmActivity
 ```
 
 
+
+## CpmProblem 
+
+```@docs
+OperationsResearchModels.CpmProblem
+```
+
+
+
+## CPM (Critical Path Method)
+
+```@docs
+OperationsResearchModels.solve(problem::CpmProblem)
+```
+
+
+
+## CpmResult 
+
+```@docs
+OperationsResearchModels.CpmResult
+```
+
+
+
+### PERT Activity
+
+```@docs
+OperationsResearchModels.PertActivity
+```
+
+
+
+### PERT Problem
+
+```@docs
+OperationsResearchModels.PertProblem
+```
+
+
+
 ## PERT (Project Evaluation and Review Technique)
+
 ```@docs
 OperationsResearchModels.solve(problem::PertProblem)
 ```
 
-### PERT Activity
+
+
+### PERT Result
+
 ```@docs
-OperationsResearchModels.PertActivity
+OperationsResearchModels.PertResult
 ```
+
