### 2025-08-28 jbytecode/OperationsResearchModels.jl#2 @dourouc05
Indeed, I think the package made a good step forward! The documentation is now much more comprehensive and self-consistent, for one. You made a good job :)! I can only see a few minor issues left.

In https://jbytecode.github.io/OperationsResearchModels.jl/dev/simplex/, the name of the function (`createsimplexproblem`) is formated in the same way as the description and the returned value. The problem is the set of spaces in the docs:

Documenter.jl interprets it as a code block (https://documenter.juliadocs.org/stable/showcase/#Code-blocks), while that's not what you mean. You don't have this formatting for the knapsack, for instance, but you have it for the TSP. It'd be worth going over the whole doc once again to find this kind of mismatch.

In a next step, you might want to use automated tools for code formatting, say JuliaFormatter.jl. Some parts are not really formatted like the others, like line lengths or use of empty lines.

The only remaining issue from my initial list, I think, is the one of determining the source and destination nodes for shortest paths. I couldn't find it in the updated documentation. (And this design choice definitely has drawbacks, which I mentioned earlier.)
