+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"scheduling/#Scheduling-Problems","page":"Scheduling","title":"Scheduling Problems","text":"","category":"section"},{"location":"scheduling/#Flow-Shop-Problem","page":"Scheduling","title":"Flow-Shop Problem","text":"","category":"section"},{"location":"scheduling/#OperationsResearchModels.Johnsons.JohnsonProblem","page":"Scheduling","title":"OperationsResearchModels.Johnsons.JohnsonProblem","text":"JohnsonProblem\n\nDescription\n\nRepresents a Johnson problem instance, containing the job processing times.\n\nFields\n\ntimes::Matrix{<:Real}: A matrix of job processing times.\n\n\n\n\n\n","category":"type"},{"location":"scheduling/#Johnson's-Rule-for-Flow-shop-Scheduling","page":"Scheduling","title":"Johnson's Rule for Flow-shop Scheduling","text":"","category":"section"},{"location":"scheduling/#OperationsResearchModels.solve-Tuple{JohnsonProblem}","page":"Scheduling","title":"OperationsResearchModels.solve","text":"johnsons(problem::JohnsonProblem)::JohnsonResult\n\nGiven a problem containing a matrix of times, returns a JohnsonResult with the permutation of the jobs.  If number of machines is 2, it uses the Johnson's algorithm for 2 machines. If number of machines is greater than 2, it uses the Johnson's algorithm by transforming the  problem into a 2-machine problem. In order to reduce the original problem to a 2-machine problem, the algorithm checks if the minimum time of the first machine is greater or equal than the maximum time of the other machines and/or if the minimum time of the  last machine is greater or equal than the maximum time of the other machines.\n\nFor example if we have 4 machines, namely, A, B, C, and D  at least one of the following conditions must be satisfied:\n\nmin(A) >= max(B, C)\nmin(D) >= max(B, C)\n\nThe function throws a JohnsonException if the problem cannot be reduced to a 2-machine problem.\n\nArguments\n\nproblem::JohnsonProblem: a problem containing a matrix of times\n\nReturns\n\nJohnsonResult: a custom data type that holds the permutation of the jobs\n\nExample\n\ntimes = Float64[\n    3.1 2.8;\n    4.0 7.0;\n    8.0 3.0;\n    5.0 8.0;\n    6.0 4.0;\n    8.0 5.0;\n    7.0 4.0\n]\n\nresult = solve(JohnsonProblem(times))\n\nprintln(result.permutation)\n\n\n\n\n\n","category":"method"},{"location":"scheduling/#Flow-Shop-Result","page":"Scheduling","title":"Flow-Shop Result","text":"","category":"section"},{"location":"scheduling/#OperationsResearchModels.Johnsons.JohnsonResult","page":"Scheduling","title":"OperationsResearchModels.Johnsons.JohnsonResult","text":"JohnsonResult\n\nDescription\n\nRepresents the result of the Johnson's algorithm, containing the optimal job permutation.\n\nFields\n\npermutation::Vector{Int}: The optimal job permutation found by the algorithm.\n\n\n\n\n\n","category":"type"},{"location":"scheduling/#Genetic-Algorithm-for-the-problems-that-cannot-be-solved-with-using-Johnson's-Rule","page":"Scheduling","title":"Genetic Algorithm for the problems that cannot be solved with using Johnson's Rule","text":"","category":"section"},{"location":"scheduling/#OperationsResearchModels.Johnsons.johnsons_ga","page":"Scheduling","title":"OperationsResearchModels.Johnsons.johnsons_ga","text":"johnsons_ga(problem::John
[…]
