### 2025-08-29 9e5b3d9
(#2) Add documentation about to heuristics used in determining the start and finish nodes
