### 2025-08-29 85a7f78
(#2) Add documentation about to heuristics used in determining the start and finish nodes in maximum flow and minimum cost flow
