+++ src/knapsack.jl
@@ -25,7 +25,7 @@ struct KnapsackResult
     selected::Vector{Bool}
     model::JuMP.Model
     objective::Float64
-end 
+end
 
 
 
@@ -86,14 +86,14 @@ function solve(problem::KnapsackProblem)::KnapsackResult
 
     n = length(values)
     model = Model(HiGHS.Optimizer)
-    
+
     MOI.set(model, MOI.Silent(), true)
 
     @variable(model, x[1:n], Bin)
     @objective(model, Max, sum(values[i] * x[i] for i = 1:n))
     @constraint(model, sum(weights[i] * x[i] for i = 1:n) <= capacity)
     optimize!(model)
-    
+
     selected = value.(x)
     return KnapsackResult(selected, model, objective_value(model))
 end
