+++ test/testtravelingsalesman.jl
@@ -1,53 +1,54 @@
-@testset "Traveling Salesman with Random Key GA" verbose = true begin 
+@testset "Traveling Salesman with Random Key GA" verbose = true begin
 
-    @testset "Points in a rectangle" begin 
-        
+    @testset "Points in a rectangle" begin
+
         #
         # * * * * * * 
         # *         *
         # * * * * * *
         pts = Float64[
-            0 0;
-            0 1;
-            0 2;
-            1 2;
-            2 2;
-            3 2;
-            4 2; 
-            5 2;
-            5 1;
-            5 0;
-            4 0;
-            3 0;
-            2 0;
-            1 0;
+            0 0
+            0 1
+            0 2
+            1 2
+            2 2
+            3 2
+            4 2
+            5 2
+            5 1
+            5 0
+            4 0
+            3 0
+            2 0
+            1 0
         ]
 
         n = size(pts, 1)
 
         distmat = zeros(n, n)
-        
+
         # create the distance matrix 
-        for i in 1:n
-            for j in 1:n
-                distmat[i, j] = sqrt(sum((pts[i, :] .- pts[j, :]).^2))
-            end 
+        for i = 1:n
+            for j = 1:n
+                distmat[i, j] = sqrt(sum((pts[i, :] .- pts[j, :]) .^ 2))
+            end
         end
 
         bestval = Inf64
 
         problem = TravelingSalesmanProblem(distmat)
 
-        for i in 1:10
-            result = solve(problem, ngen = 1000, popsize = 100, pcross = 1.0, pmutate = 0.10)
+        for i = 1:10
+            result =
+                solve(problem, ngen = 1000, popsize = 100, pcross = 1.0, pmutate = 0.10)
             if result.cost < bestval
                 bestval = result.cost
                 bestresult = result
             end
-        end 
+        end
 
         @test bestval == 14.0
-    end 
+    end
 
 
 end # end of whole @testset
