+++ test/testtransportation.jl
@@ -81,7 +81,8 @@
         t = TransportationProblem(
             Float64[10 20 25; 12 15 7],
             Float64[190, 40, 20],
-            Float64[150, 100])
+            Float64[150, 100],
+        )
 
         result = solve(t)
 
@@ -143,12 +144,12 @@
             t = TransportationProblem(
                 Float64[1 2 8 10; 7 3 4 5; 12 11 9 6],
                 Float64[50, 60, 30, 130],
-                Float64[90, 110, 70]
+                Float64[90, 110, 70],
             )
             result = leastcost(t)
             expected = Float64[
-                50 40 0 0;
-                0 20 30 60;
+                50 40 0 0
+                0 20 30 60
                 0 0 0 70
             ]
             @test result.solution == expected
@@ -159,15 +160,15 @@
             t = TransportationProblem(
                 Float64[10 2 20 11; 12 7 9 20; 4 14 16 18],
                 Float64[5, 15, 15, 15],
-                Float64[15, 25, 10]
+                Float64[15, 25, 10],
             )
-            
+
             result = leastcost(t)
-            
+
             expected = Float64[0 15 0 0; 0 0 15 10; 5 0 0 5]
-            
+
             @test result.solution == expected
-            
+
             @test result.cost == 475
         end
     end
@@ -175,7 +176,8 @@
 
 
 
-    @testset "Big Example with Nortwest-corner initial solution - Check if it is solved in reasonable times" verbose = true begin
+    @testset "Big Example with Nortwest-corner initial solution - Check if it is solved in reasonable times" verbose =
+        true begin
 
         t = TransportationProblem(
             rand(10:1000, 350, 450) * 1.0,
@@ -190,7 +192,8 @@
 
     end
 
-    @testset "Big Example with least cost initial solution - Check if it is solved in reasonable times" verbose = true begin
+    @testset "Big Example with least cost initial solution - Check if it is solved in reasonable times" verbose =
+        true begin
 
         t = TransportationProblem(
             rand(10:1000, 350, 450) * 1.0,
