### 2025-08-29 b2cbc04
(#2) Minor fixes in docstrings
