### 2025-08-29 jbytecode/OperationsResearchModels.jl#2 @jbytecode
@dourouc05 - I'm so sorry for taking your time twice. I hope it is clear now. My fixes include

- Indentation of docstrings. The documentation looks good now.
- JuliaFormatter has been applied to src and tests folders.
- I added an information for the network models that have start and finish nodes (shortest path, maximum flow, minimum cost flow). I was always suspecting on my heuristic approach, but I didn't make it ready for general use because the textbooks cover the basic cases but I plan to implement the problem structures as they allow users to specify start and finish nodes manually as well. Especially, network models with more than one start and finish nodes are not rare. I added this:

> In this function it's assumed that the problem has a unique start and finish node. A heuristic approach is used to find the start and finish nodes. If a node has only outcoming connections, it is considered the start node. If a node has only incoming connections, it is considered the finish node. Of course a network can have multiple start and finish nodes, but this heuristic simplifies the problem. Future implementations could explore more complex scenarios with multiple start and finish nodes.

I used a special formatting for this. Link:

https://jbytecode.github.io/OperationsResearchModels.jl/dev/network/#Determining-start-and-finish-nodes-6212604462e31c36

If it is okay to you, I plan to cover further improvements in the future but not in this release.

I'm really appreciated for your work and effort.

Thank you in advance.
