### 2025-08-29 c465cb2
fix nortwestcorner loop
