### 2025-09-01 jbytecode/OperationsResearchModels.jl#2 @dourouc05
It's totally fine to have more changes for future releases :).
