### 2025-09-01 #8592 @dourouc05
LGTM
