### 2025-09-01 8993211
fix minor changes in manusript
