### 2025-09-01 9e23275
fix minor changes in manusript
