### 2025-09-01 1c8fd47
biased key -> random key ga
