### 2025-09-01 #8592 @jbytecode
@check references
