### 2025-09-01 #8592 @jbytecode
@arfon - Here is the list of changes:

- The version is v0.2.7
- The archive DOI is 10.5281/zenodo.17019015
- I made minor changes in paper and fixed capitalization stuff in bibliography.
- The new release has also been registered in the Julia package registry: https://juliahub.com/ui/Packages/General/OperationsResearchModels
